requirement_ro.txt
beautifulsoup4==4.6.3
datascience==0.17.5
earthengine-api==0.1.332
easydict==1.10
easyocr==1.6.2
numpy==1.21.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
pandas==1.3.5
pandas-datareader==0.9.0
pandas-gbq==0.17.9
pandas-profiling==1.4.1
regex==2022.6.2
requests==2.23.0
requests-oauthlib==1.3.1
Pillow==7.1.2